added ParallelFileWrite and ParallelFileRead. With those functions, you can write and read BinaryBuffers in and out of one big file. Note that this is done with MPI-File I/O, so it DOES SCALE. NOTE: It DOES NOT scale to write data into files from numCores > 10000 using OTHER methods (like e.g. opening a file named myFile..GetProcRank() on each core). See also parallel_archive.cpp for functions to write a .ar-zipped big file which you can decompose into single files (like for .vtu files).

+++ ugbase/pcl/parallel_archive.cpp
@@ -59,7 +59,7 @@ void WriteParallelArchive(ProcessCommunicator &pc, std::string strFilename, cons
 
 	MPI_File_seek(fh, myOffset, MPI_SEEK_SET);
 
-	std::vector<size_t> allOffsets;
+	std::vector<int> allOffsets;
 	if(bFirst) allOffsets.resize(pc.size());
 	else allOffsets.resize(1);
 	MPI_Gather(&myOffset, 1, MPI_INT,
